Add telemetry for toolwindow editor actions
