+++ src/main/java/ee/carlrobert/codegpt/actions/ActionType.java
@@ -8,5 +8,10 @@ public enum ActionType {
   DELETE_CONVERSATION,
   DISCARD_TOKEN_LIMIT,
   OPEN_CONVERSATION_IN_EDITOR,
+  DIFF_CODE,
+  EDIT_CODE,
+  CREATE_NEW_FILE,
+  COPY_CODE,
+  REPLACE_IN_MAIN_EDITOR,
   RELOAD_MESSAGE
 }
