+++ src/main/resources/messages/codegpt.properties
@@ -58,3 +58,15 @@ dialog.tokenLimitExceeded.title=Token Limit Exceeded
 dialog.tokenLimitExceeded.description=The maximum default token limit has been reached. Do you want to proceed with the conversation despite the higher messaging cost?
 dialog.tokenLimitExceeded.continue=Continue
 dialog.tokenLimitExceeded.cancel=Cancel
+editor.diff.title=CodeGPT Diff
+editor.diff.local.content.title=CodeGPT suggested code
+toolwindow.chat.editor.action.copy.title=Copy
+toolwindow.chat.editor.action.copy.description=Copy generated code
+toolwindow.chat.editor.action.diff.title=Diff
+toolwindow.chat.editor.action.diff.description=Diff editor code against the generated one
+toolwindow.chat.editor.action.edit.title=Edit Source
+toolwindow.chat.editor.action.edit.description=Edit generated code
+toolwindow.chat.editor.action.newFile.title=New File
+toolwindow.chat.editor.action.newFile.description=Create new file from generated code
+toolwindow.chat.editor.action.replaceSelection.title=Replace Selection
+toolwindow.chat.editor.action.replaceSelection.description=Replace main editor selected code
